Moves Update/Line code into modules
This requires changing a few fields and functions on the structs to be
publicly visible, but this shouldn't really matter to us.

This _should_ make refactoring to the new backend protocols more sane
going forward.
